Breadth First Search
Breadth first traversal is a layer by layer traversal without visiting
the same node twice. In this program we do this by initialising a
queue which stores the adjacent nodes of the visited node.
completed the code for breadth first search. Initially the algorithm
adds nodes to the graph using the add edge function. This function
creates an array of lists whose key is a node and the value is the value of the neighbour.

it iterates through the nodes and goes on adding the neighbours to th
e array.