Depth First Algorithm
Graph: the creator function simply takes in the number of nodes.
it initializes the adj list which is an array of lists which cont
ain adjacent nodes of each node.
add Edge: It takes in two integers, it will add the second node as
the adjacent node of the first node.

DFSUtil is a a code simplification of DFS function. DFS util takes in
the visited array, and the node to be visited(or checked). It then
visits that node and checks its adjacent nodes by calling the DFSUtil
on the visited array and the node number.

DFS function is the fairly simple function which initialises a visite
d array and passes it into DFSUtil function.