binary search
an algorithm which searches for the given key in a sorted array
it performs the search in O(log(n)) complexity.

Note to self:seems to return negative for almost all of the searches,
need to figure out why.

Explanation: the binary search which was recursievely called in the
function was not returned and hence lead to the function ending up at
the -1 mark all the time.