hacker earth
there was a question, on hacker earth for honeywell hiring. It is
a simple question and required a very effecient method to swap an
entire array. There is this array which is first given to the pro
gram. the program is then given n number of queries. Each query
gives a range. Within this range of index numbers have to be swapped
with those numbers which are "related" to them. the relation ship being
defined as [n]-[-n].