ternaray search
ternaray search is the process of searching the position of a numbe
r in the sorted array. This can be done only on a sorted array. It
basically splits the array into three parts, (min,max1) (max1,max2)
(max2,max). The number is searched in the first second and third
part. The program will return the key if such a key exist otherwise
it will return -1.