rowcol count
the problem was with initialization of column count. this was to be
inted to 1. this proved to be tricky because the problem became app
arent only after i saw the output of what happens when the resulting
max count is merely incremented by 1. with that however the problem is
done.