merge array
the program, is to take in put from a text file and output to stdio
which will also be taken by a text file. added one more element to
make the arrays equal. somehow the program has a problem when two
values of equal magnitude are present in both arrays.