benfords law
adding a program to track benfords law in documents. benfords law
states that just like languages, numbers have a language of their
own. They can be examined to see if the distribution matches the
actual occurence. if it doesn't then its probable that someone is
trying to dupe you. added a small correction for matching with type.
added zero initialiser for arrays.